Move shared methods out to helpers module
Includes removal of `version` attr_accessor from
gem publisher. This was left over from a previous
version where it was called on the instance in one
of the rake tasks.
