+++ build_tools/publisher/gem_publisher.rb
@@ -1,10 +1,11 @@
 require 'govuk_template/version'
+require_relative '../helpers'
 require 'tmpdir'
 require 'open3'
 
 module Publisher
   class GemPublisher
-    attr_accessor :version
+    include Helpers
 
     GIT_URL = "https://#{ENV['GITHUB_TOKEN']}@github.com/alphagov/govuk_template.git"
 
@@ -22,19 +23,6 @@ module Publisher
           run "git push --tags origin master"
         end
       end
-    end
-
-    def version_released?
-      output = run "git ls-remote --tags #{GIT_URL.shellescape}"
-      return !! output.match(/v#{@version}/)
-    end
-
-  private
-
-    def run(command)
-      output, status = Open3.capture2e(command)
-      abort "Error running #{command}: exit #{status.exitstatus}\n#{output}" if status.exitstatus > 0
-      output
     end
   end
 end
