+++ build_tools/publisher/jinja_publisher.rb
@@ -1,9 +1,11 @@
 require 'govuk_template/version'
+require_relative '../helpers'
 require 'tmpdir'
 require 'open3'
 
 module Publisher
   class JinjaPublisher
+    include Helpers
     GIT_URL = "https://#{ENV['GITHUB_TOKEN']}@github.com/alphagov/govuk_template_jinja.git"
 
     def initialize(version = GovukTemplate::VERSION)
@@ -25,19 +27,6 @@ module Publisher
           run "npm publish ./"
         end
       end
-    end
-
-    def version_released?
-      output = run("git ls-remote --tags #{GIT_URL.shellescape}")
-      return !! output.match(/v#{@version}/)
-    end
-
-  private
-
-    def run(command)
-      output, status = Open3.capture2e(command)
-      abort "Error running #{command}: exit #{status.exitstatus}\n#{output}" if status.exitstatus > 0
-      output
     end
   end
 end
